unification of the dankness
