--- barbaric/info.json
@@ -1,8 +1,0 @@
-{
-    "AUTHOR" : "UltimatePancake",
-    "INSTALL_MSG" : "[p]barbaric to go dank.",
-    "NAME" : "Barbaric",
-    "SHORT" : "Absolutely barbaric",
-    "DESCRIPTION" : "Displays the 'absolutely barbaric' meme.",
-    "TAGS ": ["meme", "barbaric"]
-}
