--- dickbutt/info.json
@@ -1,8 +1,0 @@
-{
-    "AUTHOR" : "UltimatePancake",
-    "INSTALL_MSG" : "[p]dickbutt to display a dickbutt",
-    "NAME" : "Dickbutt",
-    "SHORT" : "It displays a dickbutt",
-    "DESCRIPTION" : "It displays a dickbutt, nothing more, nothing less.",
-    "TAGS" : ["dickbutt", "meme", "dank"]
-}
